Wipe ChillDontKill2
It had a finish but it was not any good.
Or at least not something that could be named ChillDontKill2
and not worth to be renamed to "uslessmap3000.map"

No front chiller but nobody would ever find use in this. If you want to
feel nostalgic in 30 years go restore it from git. But this is only
polluting my mapdir.